requirements.txt
# Core
fastapi>=0.68.1,<1.0.0
uvicorn[standard]>=0.15.0,<1.0.0
python-dotenv>=0.19.0,<1.0.0
python-multipart>=0.0.5,<1.0.0

# Data Processing
pandas>=1.3.3,<2.0.0
numpy>=1.21.2,<2.0.0
scipy>=1.7.0,<2.0.0

# Machine Learning
scikit-learn>=1.0.2,<2.0.0
tensorflow>=2.9.0,<3.0.0
torch>=1.10.0,<2.0.0
torchvision>=0.11.0,<0.16.0
joblib>=1.1.0,<2.0.0
statsmodels>=0.13.0,<1.0.0

# Computer Vision
opencv-python-headless>=4.5.4,<5.0.0
Pillow>=9.0.0,<10.0.0

# NLP
transformers>=4.15.0,<5.0.0
nltk>=3.6.0,<4.0.0

# Time Series
prophet>=1.0.0,<2.0.0
pystan>=2.19.1,<3.0.0

# Visualization
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.2,<0.13.0
plotly>=5.3.0,<6.0.0

# Utilities
tqdm>=4.41.0,<5.0.0
pyyaml>=5.4.1,<7.0.0
python-json-logger>=2.0.0,<3.0.0

# API
gunicorn>=20.1.0,<21.0.0
httpx>=0.23.0,<0.24.0
pydantic>=1.9.0,<2.0.0
